requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
_pytorch_select=0.1=cpu_0
absl-py=1.0.0=pyhd8ed1ab_0
aiohttp=3.8.1=py39h7f8727e_0
aiosignal=1.2.0=pyhd8ed1ab_0
altgraph=0.17=pyhd3eb1b0_0
async-timeout=4.0.1=pyhd8ed1ab_0
attrs=21.2.0=pyhd8ed1ab_0
blas=1.0=mkl
blinker=1.4=py_1
blosc=1.21.0=h8c45485_0
bottleneck=1.3.2=py39hdd57654_1
brotli=1.0.9=he6710b0_2
brotlipy=0.7.0=py39h3811e60_1001
brunsli=0.1=h2531618_0
bzip2=1.0.8=h7b6447c_0
c-ares=1.17.1=h27cfd23_0
ca-certificates=2021.10.26=h06a4308_2
cachetools=4.2.4=pyhd8ed1ab_0
certifi=2021.10.8=py39h06a4308_0
cffi=1.14.6=py39h400218f_0
cfitsio=3.470=hf0d0db6_6
charls=2.2.0=h2531618_0
charset-normalizer=2.0.9=pyhd8ed1ab_0
click=7.1.2=pypi_0
cloudpickle=2.0.0=pyhd3eb1b0_0
colorama=0.4.4=pyh9f0ad1d_0
cryptography=3.4.8=py39hbca0aa6_0
cycler=0.10.0=py39h06a4308_0
cytoolz=0.11.0=py39h27cfd23_0
dask-core=2021.10.0=pyhd3eb1b0_0
dataclasses=0.8=pyhc8e2a94_3
dbus=1.13.18=hb2f20db_0
expat=2.4.1=h2531618_2
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=h70c0345_0
frozenlist=1.2.0=py39h7f8727e_0
fsspec=2021.11.1=pyhd8ed1ab_0
future=0.18.2=py39hf3d152e_4
giflib=5.2.1=h7b6447c_0
glib=2.69.1=h5202010_0
gmp=6.2.1=h2531618_2
google-auth=2.3.3=pyh6c4a22f_0
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
grpcio=1.42.0=py39hce63b2e_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
icu=58.2=he6710b0_3
idna=3.1=pyhd3deb0d_0
imagecodecs=2021.8.26=py39h4cda21f_0
imageio=2.9.0=pyhd3eb1b0_0
importlib-metadata=4.8.2=py39hf3d152e_0
intel-openmp=2019.4=243
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9d=h7f8727e_0
jxrlib=1.1=h7b6447c_2
kiwisolver=1.3.1=py39h2531618_0
krb5=1.19.2=hac12032_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
lerc=3.0=h295c915_0
libaec=1.0.4=he6710b0_1
libblas=3.9.0=1_h6e990d7_netlib
libcblas=3.9.0=3_h893e4fe_netlib
libcurl=7.80.0=h0b77cf5_0
libdeflate=1.8=h7f8727e_5
libedit=3.1.20210910=h7f8727e_0
libev=4.33=h7f8727e_1
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
liblapack=3.9.0=3_h893e4fe_netlib
libmklml=2019.0.5=0
libnghttp2=1.46.0=hce63b2e_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.15.8=h780b84a_0
libssh2=1.9.0=h1ba5d50_1
libstdcxx-ng=9.3.0=hd4cf53a_17
libtiff=4.2.0=h85742a9_0
libuuid=1.0.3=h7f8727e_2
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
libzopfli=1.0.3=he6710b0_0
locket=0.2.1=py39h06a4308_1
lz4-c=1.9.3=h295c915_1
macholib=1.14=pyhd3eb1b0_1
markdown=3.3.6=pyhd8ed1ab_0
matplotlib=3.4.3=py39h06a4308_0
matplotlib-base=3.4.3=py39hbbc1b5f_0
mkl=2020.2=256
multidict=5.1.0=py39h27cfd23_2
munkres=1.1.4=py_0
ncurses=6.3=h7f8727e_2
networkx=2.6.3=pyhd3eb1b0_0
ninja=1.10.2=h4bd325d_0
numpy=1.20.3=py39hdbf815f_1
oauthlib=3.1.1=pyhd8ed1ab_0
olefile=0.46=pyhd3eb1b0_0
openjpeg=2.4.0=h3ad879b_0
openssl=1.1.1l=h7f8727e_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.2.5=py39hde0f152_0
partd=1.2.0=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pillow=8.4.0=py39h5aabda8_0
pip=21.2.4=py39h06a4308_0
protobuf=3.15.8=py39he80948d_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.21=pyhd8ed1ab_0
pycryptodome=3.10.1=py39h3dc18e1_0
pydeprecate=0.3.1=pyhd8ed1ab_0
pyinstaller=3.6=py39h8f29fe8_6
pyjwt=2.3.0=pyhd8ed1ab_1
pyopenssl=21.0.0=pyhd8ed1ab_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py39h2531618_6
pyqt5=5.15.4=pypi_0
pyqt5-plugins=5.15.4.2.2=pypi_0
pyqt5-qt5=5.15.2=pypi_0
pyqt5-sip=12.9.0=pypi_0
pyqt5-tools=5.15.4.3.2=pypi_0
pysocks=1.7.1=py39hf3d152e_4
python=3.9.7=h12debd9_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python-dotenv=0.19.2=pypi_0
python_abi=3.9=2_cp39
pytorch=1.8.1=cpu_py39h60491be_0
pytorch-lightning=1.5.6=pyhd8ed1ab_0
pytz=2021.3=pyhd3eb1b0_0
pyu2f=0.1.5=pyhd8ed1ab_0
pywavelets=1.1.1=py39h6323ea4_4
pyyaml=5.4.1=py39h3811e60_0
qt=5.9.7=h5867ecd_1
qt5-applications=5.15.2.2.2=pypi_0
qt5-tools=5.15.2.1.2=pypi_0
readline=8.1=h27cfd23_0
requests=2.26.0=pyhd8ed1ab_1
requests-oauthlib=1.3.0=pyh9f0ad1d_0
rsa=4.8=pyhd8ed1ab_0
scikit-image=0.18.3=py39h51133e4_0
scikit-learn=1.0.1=py39h51133e4_0
scipy=1.5.3=py39hf3f25e7_0
setuptools=58.0.4=py39h06a4308_0
sip=4.19.13=py39h2531618_0
six=1.16.0=pyhd3eb1b0_0
snappy=1.1.8=he6710b0_0
sqlite=3.36.0=hc218d9a_0
tensorboard=2.7.0=pyhd8ed1ab_0
tensorboard-data-server=0.6.0=py39h3da14fd_0
tensorboard-plugin-wit=1.8.0=pyh44b312d_0
threadpoolctl=2.2.0=pyh0d69192_0
tifffile=2021.7.2=pyhd3eb1b0_2
tk=8.6.11=h1ccaba5_0
toolz=0.11.2=pyhd3eb1b0_0
torchmetrics=0.6.2=pyhd8ed1ab_0
torchvision=0.2.2=py_3
tornado=6.1=py39h27cfd23_0
tqdm=4.62.3=pyhd8ed1ab_0
typing-extensions=4.0.1=hd8ed1ab_0
typing_extensions=4.0.1=pyha770c72_0
tzdata=2021e=hda174b7_0
urllib3=1.26.7=pyhd8ed1ab_0
werkzeug=2.0.1=pyhd8ed1ab_0
wheel=0.37.0=pyhd3eb1b0_1
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h516909a_0
yarl=1.6.3=py39h3811e60_2
zfp=0.5.5=h2531618_6
zipp=3.6.0=pyhd8ed1ab_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
